add single-project.tsx + refact projects-section.tsx
